% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{drop_levels}
\alias{drop_levels}
\title{drop_levels}
\usage{
drop_levels(x, ...)
}
\arguments{
\item{x}{an object from which to drop unsued factor levels.}

\item{...}{further arguments passed to base droplevels.}
}
\description{
The function drop_levels is used to drop unused levels from a factor or, more
commonly, from factors in a data frame. Unlike the base function, drop_levels
will preserve labels.
}
\seealso{
\code{\link[base]{droplevels}}
}
