% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{droplevels}
\alias{droplevels}
\title{droplevels}
\usage{
droplevels(x, ...)
}
\arguments{
\item{x}{an object from which to drop unsued factor levels.}
}
\description{
The function dropevels is used to drop unused levels from a factor or, more
commonly, from factors in a data frame. Unlike the base function, droplevels
will preserve labels.
}
\seealso{
\code{\link[base]{droplevels}}
}
