% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-extras.R
\name{fct_drop_levels}
\alias{fct_drop_levels}
\title{Drop unused levels}
\usage{
fct_drop_levels(f, only)
}
\arguments{
\item{f}{A factor}

\item{only}{A character vector restricting the set of levels to be dropped.
If supplied, only levels that have no entries and appear in this vector
will be removed.}
}
\description{
Wrapper around forcats drop_levels which will not modify or drop
non-level attributes. Compared to `base::droplevels()` does not drop `NA`
levels that have values.
}
\examples{
f <- factor(c("a", "b"), levels = c("a", "b", "c"))
f
fct_drop_levels(f)

# Set only to restrict which levels to drop
fct_drop_levels(f, only = "a")
fct_drop_levels(f, only = "c")
}
\seealso{
\code{\link[forcats]{fct_drop}}
}
