% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{\%+\%}
\alias{\%+\%}
\title{\%+\%}
\usage{
x \%+\% y
}
\arguments{
\item{x}{The first vector to call the operator on.}

\item{y}{The second vector to call the operator on.}
}
\description{
Allows concatenation of character vectors. A short form of basic paste0
functionality.
}
\seealso{
\code{\link[base]{Arithmetic}}
}
