% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ifelse}
\alias{ifelse}
\title{ifelse}
\usage{
ifelse(test, yes, no, attributes = "default")
}
\arguments{
\item{test}{an object which can be coerced to logical mode.}

\item{yes}{return values for true elements of test.}

\item{no}{return values for false elements of test.}

\item{attributes}{Whether to optionally preserve attributes, specifically
levels and class. Preserved attributes can be "default", which mimics
the default base ifelse behaviour (and is the default), or to preserve
the attributes of either "test", "yes" or "no".}
}
\description{
ifelse returns a value with the same shape as test which is filled with
elements selected from either yes or no depending on whether the element of
test i or FALSE. Unlike the base function, ifelse optionally allows
preservation of attributes, either from the yes or no parameters.
}
\seealso{
\code{\link[base]{ifelse}}
}
