% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{key_value}
\alias{key_value}
\title{key_value}
\usage{
key_value(lst)
}
\arguments{
\item{lst}{A named list or vector.}
}
\value{
A list with name elements stored as $key and values stored as $value
        for each original element in the vector or list.
}
\description{
Return a list or vector with names becoming $key elements and value
elements becoming $value elements.
}
\examples{
vect <- list('A', 'B', 'C')
names(vect) <- c('X', 'Y', 'Z')
for (el in key_value(vect)) {
  print(paste(el$key, el$value))
}

}
