% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{levelNA}
\alias{levelNA}
\title{NA to factor level}
\usage{
levelNA(df, missingName = "Missing", checkNA = TRUE, except = NULL)
}
\arguments{
\item{df}{Data frame to change the levels to.}

\item{missingName}{What to call the new NA level.}

\item{checkNA}{Add a level only if there are NA factors in a given factor.}

\item{except}{A regex of column names not to match.}
}
\value{
data.frame
}
\description{
Change NA to a factor level in data frame factor variables.
}
\examples{
df <- data.frame(a = factor(c('a', 'b', NA)), b = factor(c('x', 'y', 'z')))

levelNA(df)
levelNA(df, missingName = 'Not Available')
levelNA(df, checkNA = FALSE)
levelNA(df, except = '^(a|b)'

}
