% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{lm2}
\alias{lm2}
\title{lm model with robust standard errors.}
\usage{
lm2(..., robust.se = TRUE)
}
\arguments{
\item{robust.se}{Whether to use robust standard errors.}
}
\description{
Estimate an lm model but uses robust standard errors by default.
}
\examples{
data(iris)
iris$setosa <- ifelse(iris$Species == 'setosa', 1, 0)
lmModel <- lm2(setosa ~ Sepal.Length + Sepal.Width, data = iris)

}

