% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{load_remote_zip}
\alias{load_remote_zip}
\title{load_remote_zip}
\usage{
load_remote_zip(remote_zip)
}
\arguments{
\item{remote_zip}{A url to a web address containing a zipped R package.}
}
\description{
Load a remote zip file as a library. Nothing is permanently installed.
While technically this _should_ work for any package, its use is in
practice reserved for packages small enough to be very quickly loaded.
}
\examples{
\dontrun{
load_remote_zip('/path/to/some_zipped_package.zip')
}

}
