% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-extras.R
\name{mutateX}
\alias{mutateX}
\title{mutateX}
\usage{
mutateX(.data, ..., .varX = "X")
}
\arguments{
\item{.data}{A tbl.}

\item{...}{Name-value pairs of expressions.}

\item{.varX}{What to use as a substitute.}
}
\value{
An object of the same class as .data.
}
\description{
mutate which replaces X with the name of the assigned column.
}
\examples{
data <- data.frame(some_var = 1:100)
mutateX(data, some_var = ifelse(X > 10, 10, X), noChange = some_var + 1)
# Or in vanilla mutate:
library(dplyr)
data <- data.frame(some_var = 1:100)
mutate(data, some_var = ifelse(some_var > 10, 10, some_var),
       noChange = some_var + 1)

}
