% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{+}
\alias{+}
\title{+}
\usage{
"+"(x, y)
}
\arguments{
\item{x}{The first vector to call the '+' operator on.}

\item{y}{The second vector to call the '+' operator on.}
}
\description{
Same as the base '+' arithmetic operator, with the exception that it
overloads it, allowing concatenation of character vectors. Note that there is
no type coercion. The function will throw an error if character is mixed with
other types.
}
\seealso{
\code{\link[base]{Arithmetic}}
}
