% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{strip_glm}
\alias{strip_glm}
\title{strip_glm}
\usage{
strip_glm(cm)
}
\arguments{
\item{cm}{The glm model to strip.}
}
\value{
Stripped glm model.
}
\description{
Strip the glm model, leaving only the necessary bits for doing a predict.
}
\examples{
data(iris)
iris$setosa <- ifelse(iris$Species == 'setosa', 1, 0)
sglm <- strip_glm(glm(setosa ~ Sepal.Length + Sepal.Width, data = iris))

}
