% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{stripGlm}
\alias{stripGlm}
\title{Strip glm model.}
\usage{
stripGlm(cm)
}
\arguments{
\item{cm}{The glm model to strip.}
}
\value{
Stripped glm model.
}
\description{
Strip the glm model, leaving only the necessary bits for doing a predict.
}
\examples{
data(iris)
iris$setosa <- ifelse(iris$Species == 'setosa', 1, 0)
sglm <- stripGlm(glm(setosa ~ Sepal.Length + Sepal.Width, data = iris))

}

