% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{robust_lm}
\alias{robust_lm}
\title{robust_lm}
\usage{
robust_lm(..., robust.se = TRUE)
}
\arguments{
\item{robust.se}{Whether to use robust standard errors.}
}
\description{
Estimate an lm model but uses robust standard errors by default.
}
\examples{
data(iris)
iris$setosa <- ifelse(iris$Species == 'setosa', 1, 0)
lm_model <- robust_lm(setosa ~ Sepal.Length + Sepal.Width, data = iris)

}
