% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{robust_lm}
\alias{robust_lm}
\title{robust_lm}
\usage{
robust_lm(...)
}
\arguments{
\item{...}{Arguments to pass to lm.}
}
\description{
Estimates an lm model but uses robust standard errors by default. Note that
for functions for which a robust_lm method is not defined (which is, in
effect all of them with the exception of broom's tidy), the fallback class is
lm, which will yield standard (non-robust) standard errors.
}
\examples{
data(iris)
iris$setosa <- ifelse(iris$Species == 'setosa', 1, 0)
lm_model <- robust_lm(setosa ~ Sepal.Length + Sepal.Width, data = iris)

}
