% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-extras.R
\name{vf}
\alias{vf}
\title{Display filtered data frame.}
\usage{
vf(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Conditions on the data frame. Comma-separated conditions are
joined by \code{&}}
}
\value{
Invisible \code{NULL}. Result is displayed in a window.
}
\description{
\code{vf} displays a data frame using the View function but allows conditions
on the data by using \code{dplyr::filter} and displays the conditions in the
windows title.
}
\examples{
vf(iris, Sepal.Width >= 4)
vf(iris, Sepal.Length > 6, Sepal.Width < 4)
}
