% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-extras.R
\name{vs}
\alias{vs}
\title{Display selected elements from a data frame.}
\usage{
vs(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Names the variables to be selected.}
}
\value{
Invisible \code{NULL}. Result is displayed in a window.
}
\description{
\code{vs} displays a data frame using the View function but allows selecting
variables from the data using \code{dplyr::select} and displays the
number of selected variables in the windows title.
}
\examples{
\dontrun{
vs(iris, Sepal.Width)
vs(iris, Sepal.Length, Sepal.Width)
}

}
